requirements.txt
numpy==1.19.5
scipy==1.6.0
Pillow==8.1.0
cython
matplotlib==3.3.4
scikit-image==0.18.1
tensorflow==2.4.1
keras==2.2.5
opencv-python==4.5.1.48
h5py
imgaug==0.4.0
IPython[all]==7.20.0
tqdm==4.56.0
pycocotools==2.0.2
uvicorn==0.13.3
fastapi==0.63.0
python-multipart
requests==2.25.1
pandas==1.2.2